Simplify inspector API and consolidate build configs into single vite config
