+++ lix/packages/inspector/vite.config.ts
@@ -1,20 +1,33 @@
+/**
+ * Special build for library mode.
+ */
 import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
+import { dirname, resolve } from "node:path";
+import { fileURLToPath } from "node:url";
+import dts from "vite-plugin-dts";
+import path from "node:path";
 import tailwindcss from "@tailwindcss/vite";
-import path from "path";
+import react from "@vitejs/plugin-react";
 
-export const plugins = [tailwindcss(), react()];
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 export default defineConfig({
-  plugins,
+  plugins: [
+    tailwindcss(),
+    react(),
+    dts({
+      include: ["src/index.tsx"],
+    }),
+  ],
   build: {
+    minify: false,
     target: "es2022",
-    minify: true,
-    outDir: "build",
-  },
-  resolve: {
-    alias: {
-      "@": path.resolve(__dirname, "./src"),
+    lib: {
+      entry: resolve(__dirname, "src/index.tsx"),
+      formats: ["es"],
+      fileName: "index",
+      cssFileName: "styles",
     },
   },
+  
 });
